src/requirements.txt
fastapi==0.70
orjson==3.6.4
uvicorn==0.15.0
uvloop==0.16.0
gunicorn==20.1.0
httptools==0.3.0
backoff==1.11.1
grpcio-tools==1.42.0
opentracing==2.4.0
jaeger-client==4.8.0
aiokafka==0.6.0
pyjwt==2.3.0
sentry-sdk==1.5.1
python-logstash==0.4.6